Mesos Check: Master State and Stats check.
Provides a mechanism to check the Apache Mesos Master State and
Stats endpoints.

State
=====

The following metrics are reported as obtained from `master/state.json` :
    * Value of "activated_slaves"
    * Number of "slaves"
    * Number of "completed_frameworks"
    * Value of "deactivated_slaves"
    * Value of "failed_tasks"
    * Value of "finished_tasks"
    * Value of "killed_tasks"
    * Value of "lost_tasks"
    * Value of "staged_tasks"
    * Value of "started_tasks"

In addition any metrics that are reported are tagged with the value of
the fields  version, leader, pid, id, git_branch, git_sha as obtained
from the endpoint. The convention of "<name of tag>:<value>" is
followed.

e.g. "version:0.17.0"

Stats
=====

All metrics are reported to Data Dog as obtained from
`master/stats.json`. The metrics are also tagged with
the tags extracted from the *State* endpoint.

[ticket: NA]: NA